Rename holdTypes and holdUtils to holdPageTypes and holdPageUtils
